Added 2015 Day 11 and fixed test results to be of type object instead.

+++ AdventOfCode/2015/AdventOfCode2015071.cs
@@ -23,7 +23,7 @@ namespace AdventOfCode._2015
                 .Select(i => i.Split(' '))
                 .ToDictionary(i => i.Last());
 
-            Result = EvalInput("a");
+            Result = (int) EvalInput("a");
         }
 
         ushort EvalInput(string wire)
